+++ test/vault/Vault.spec.ts
@@ -96,10 +96,16 @@ describe("Vault spec", () => {
             expect(await usdc.balanceOf(vault.address)).to.eq(parseUnits("100", await usdc.decimals()))
 
             // update sender's balance
-            expect(await vault.balanceOf(alice.address)).to.eq(parseUnits("100", await usdc.decimals()))
+            expect(await vault.getBalance(alice.address)).to.eq(parseUnits("100", await usdc.decimals()))
         })
 
-        it.skip("non-standard ERC20 (USDT) is supported")
+        it("force error, inconsistent vault balance with deflationary token", async () => {
+            usdc.setTransferFeeRatio(50)
+            await expect(
+                vault.connect(alice).deposit(usdc.address, parseUnits("100", await usdc.decimals())),
+            ).to.be.revertedWith("V_IBA")
+            usdc.setTransferFeeRatio(0)
+        })
 
         it("can't add collateral not supported")
     })
@@ -136,7 +142,7 @@ describe("Vault spec", () => {
             expect(balanceAfter.sub(balanceBefore)).to.eq(amount)
 
             // update sender's balance in vault
-            expect(await vault.balanceOf(alice.address)).to.eq("0")
+            expect(await vault.getBalance(alice.address)).to.eq("0")
         })
 
         it("force error if the freeCollateral is not enough", async () => {
@@ -154,9 +160,9 @@ describe("Vault spec", () => {
     })
 
     // TODO not sure if this is needed
-    describe("balanceOf", () => {
+    describe("getBalance", () => {
         it("is zero by default", async () => {
-            expect(await vault.balanceOf(alice.address)).eq(0)
+            expect(await vault.getBalance(alice.address)).eq(0)
         })
 
         it("equals to USDC balance if it's the only collateral")